refactor(rust): rename mod layer to frames
