+++ rust/mlt-core/src/frames/mod.rs
@@ -10,8 +10,8 @@ use borrowme::borrowme;
 use integer_encoding::VarIntWriter as _;
 use utils::BinarySerializer as _;
 
-use crate::layer::unknown::Unknown;
-use crate::layer::v01::Layer01;
+use crate::frames::unknown::Unknown;
+use crate::frames::v01::Layer01;
 use crate::utils::{checked_sum2, parse_u8, parse_varint, take};
 use crate::{MltError, MltRefResult, utils};
