+++ rust/mlt-core/src/frames/optimizer.rs
