+++ rust/mlt-core/src/frames/v01/geometry/encode.rs
